+++ Assets/AppCenter/Plugins/AppCenterSDK/Analytics/Blank/AnalyticsInternal.cs
@@ -78,6 +78,15 @@ namespace Microsoft.AppCenter.Unity.Analytics.Internal
         public static void Resume()
         {
         }
+
+        public static void EnableManualSessionTracker()
+        {
+        }
+
+        public static void StartSession()
+        {
+        }
+
     }
 }
 #endif