+++ Assets/AppCenter/Plugins/AppCenterSDK/Analytics/Shared/Analytics.cs
@@ -116,5 +116,16 @@ namespace Microsoft.AppCenter.Unity.Analytics
         {
             AnalyticsInternal.Resume();
         }
+
+
+        public static void EnableManualSessionTracker()
+        {
+            AnalyticsInternal.EnableManualSessionTracker();
+        }
+
+        public static void StartSession()
+        {
+            AnalyticsInternal.StartSession();
+        }
     }
 }