+++ Assets/AppCenter/Plugins/AppCenterSDK/Analytics/iOS/AnalyticsInternal.cs
@@ -79,6 +79,16 @@ namespace Microsoft.AppCenter.Unity.Analytics.Internal
             appcenter_unity_analytics_resume();
         }
 
+        public static void EnableManualSessionTracker()
+        {
+            appcenter_unity_analytics_enable_manual_session_tracker();
+        }
+
+        public static void StartSession()
+        {
+            appcenter_unity_analytics_start_session();
+        }
+
 #region External
 
         [DllImport("__Internal")]